Implement json hot reloading
