+++ src/main/kotlin/bountiful/registry/ValueRegistry.kt
@@ -12,4 +12,8 @@ open class ValueRegistry {
         }
     }
 
+    fun empty() {
+        items.clear()
+    }
+
 }
